# Vigen-Repository/ClientApp/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Vigen-Repository",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost:44385"
    }
  ],
  "paths": {
...
    "/api/Organization/{id}": {
      "get": {
        "tags": [
          "Organization"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Organization"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Organization"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ViolenceType/{id}": {
      "get": {
        "tags": [
          "ViolenceType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceType"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ViolenceType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViolenceType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ViolenceType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ViolenceType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceType"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ViolenceType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceType"
                }
              }
            }
          }
        }
      }
    },
    "/api/ViolenceTypeOrganization": {
      "get": {
        "tags": [
          "ViolenceTypeOrganization"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ViolenceTypesOrganization"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ViolenceTypesOrganization"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ViolenceTypesOrganization"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ViolenceTypeOrganization"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViolenceTypesOrganization"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ViolenceTypesOrganization"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ViolenceTypesOrganization"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              }
            }
          }
        }
      }
    },
    "/api/ViolenceTypeOrganization/{id}": {
      "get": {
        "tags": [
          "ViolenceTypeOrganization"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ViolenceTypeOrganization"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViolenceTypesOrganization"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ViolenceTypesOrganization"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ViolenceTypesOrganization"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ViolenceTypeOrganization"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolenceTypesOrganization"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Organization": {
        "type": "object",
        "properties": {
          "nit": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "tel": {
            "type": "string",
            "nullable": true
          },
          "sites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Site"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Site": {
        "type": "object",
        "properties": {
          "nit": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "ubication": {
            "type": "string",
            "nullable": true
          },
          "range": {
            "type": "integer",
            "format": "int32"
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "tel": {
            "type": "string",
            "nullable": true
          },
          "nitNavigation": {
            "$ref": "#/components/schemas/Organization"
          }
        },
        "additionalProperties": false
      },
...
      "ViolenceType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ViolenceTypesOrganization": {
        "type": "object",
        "properties": {
          "nit": {
            "type": "string",
            "nullable": true
          },
          "idViolence": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}